requirements.txt
fastapi
uvicorn
gunicorn
python-multipart
jinja2
python-jose[cryptography]
aiofiles
python-dotenv
sqlalchemy
aiosqlite
httpx
